♻️ refactor(docs-kit): dedupe legacy-redirect types and drop hardcoded site host
Move the documentation-inventory shape (DemoOptions/DemoReference/
DocumentRecord/DocumentationInventory) into a single canonical location at
packages/docs-kit/src/types.ts, imported by both docs.config.ts and the
compiler layer, removing the structurally duplicated Legacy* types and their
`as DocumentationInventory` casts in finalizeBuild.ts and manifests.ts.

Also replace the hardcoded ui.lobehub.com PAGEFIND_BASE_URL in
pagefindEngine.ts with siteConfig.siteUrl sourced from the
virtual:lobedocs/site-config module, so the kit no longer bakes in
consumer-specific hostname data.
