+++ packages/docs-kit/site/compiler/manifests.ts
@@ -3,16 +3,15 @@ import path from 'node:path';
 import { getDocsConfig } from '../../src/config';
 import { createContentManifest } from './content/createManifest';
 import { getStandaloneDemoPaths } from './demo/readLegacyMap';
-import type { DocumentationInventory } from './types';
 
 const repositoryRoot = path.resolve(import.meta.dirname, '../../../..');
 
 export function getPrerenderPaths(): string[] {
   const config = getDocsConfig(repositoryRoot);
-  const legacyRedirects = (config.legacyRedirects ?? {
+  const legacyRedirects = config.legacyRedirects ?? {
     demoReferences: [],
     documents: [],
-  }) as DocumentationInventory;
+  };
 
   return [
     ...createContentManifest(repositoryRoot).documents.map(({ pathname }) => pathname),
