+++ packages/docs-kit/site/compiler/types.ts
@@ -1,29 +1,6 @@
-export interface DemoOptions {
-  inline: boolean;
-  isolated: boolean;
-  layout: 'default' | 'center' | 'bare';
-}
-
-export interface DemoReference {
-  document: string;
-  legacyId: string;
-  legacyRouteId: string;
-  options: DemoOptions;
-  pathname: string;
-  source: string;
-}
-
-export interface DocumentRecord {
-  category?: string;
-  description?: string;
-  legacyRouteId: string;
-  pathname: string;
-  section: string;
-  source: string;
-  title?: string;
-}
-
-export interface DocumentationInventory {
-  demoReferences: DemoReference[];
-  documents: DocumentRecord[];
-}
+export type {
+  DemoOptions,
+  DemoReference,
+  DocumentationInventory,
+  DocumentRecord,
+} from '../../src/types';
